requirements.txt
pytest
pytest-bdd==3.4.0
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.6.3-cp39-cp39-win_amd64.whl
allure-pytest